fix: Hermes crash on Android
