--- android/src/main/java/com/facebook/react/views/textinput/SelectionWatcher.kt
@@ -1,7 +1,0 @@
-// re-declaration of SelectionWatcher from react-native, but with public visibility
-
-package com.facebook.react.views.textinput
-
-interface SelectionWatcher {
-  fun onSelectionChanged(start: Int, end: Int)
-}
